[UBA] * Fixes to make macos work properly with remote helpers
[CL 30280233 by henrik karlsson in ue5-main branch]

+++ Engine/Source/Programs/UnrealBuildAccelerator/Common/Private/UbaSession.cpp
@@ -1020,9 +1020,12 @@ namespace uba
 					if (seconds <= info.deleteSessionsOlderThanSeconds)
 						return;
 
-					StringBuffer<> sessionDir(sessionsDir);
-					sessionDir.EnsureEndsWithSlash().Append(e.name);
-					DeleteAllFiles(m_logger, sessionDir.data);
+					if (IsDirectory(e.attributes)) // on macos we get a ".ds_store" file created by the os
+					{
+						StringBuffer<> sessionDir(sessionsDir);
+						sessionDir.EnsureEndsWithSlash().Append(e.name);
+						DeleteAllFiles(m_logger, sessionDir.data);
+					}
 				});
 		}
